refactor: restructure Copy.swift into organized modules
- Split Copy.swift into three focused files:
  - Copy.swift: UI text elements (buttons, labels, tabs, links)
  - Labels.swift: section titles and headings
  - Messages.swift: user messages and poetic text
- Update references from Copy.Timer to Labels.Timer
- Maintain backward compatibility for existing Copy.Button, Copy.Label, etc.
- Improve code organization and maintainability
